requirements.txt
absl-py==2.2.0
asgiref==3.8.1
astunparse==1.6.3
build==1.2.2.post1
cachetools==5.5.2
certifi==2022.9.24
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.0
CurrencyConverter==0.14.2
cycler==0.12.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.50.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-embed-video==1.4.8
django-environ==0.4.5
django-heroku==0.3.1
django-paypal==1.0.0
django-storages==1.9.1
ez_setup==0.9
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==20.0.4
h5py==3.13.0
idna==2.9
keras==3.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mysqlclient==1.4.6
namex==0.0.8
numpy==2.1.3
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
paytm==0.1.8
pillow==11.1.0
protobuf==5.29.4
psycopg2==2.9.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pycrypto==2.6.1
pycryptodome==3.9.8
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2022.6
requests==2.23.0
requests-oauthlib==1.3.1
rich==13.9.4
rsa==4.0
secure-smtplib==0.1.1
six==1.16.0
sqlparse==0.3.1
stripe==2.48.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf-estimator-nightly==2.8.0.dev2021122109
tf-keras==2.14.1
tomli==2.2.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==1.25.9
Werkzeug==3.1.3
whitenoise==5.2.0
wrapt==1.17.2
